Fix crash caused by autocomplete-min-length=0
The computed property `autoCompleteOpen` checks that the list
`filteredAutocompleteItems` is not empty. The issue is, to compute that
list we check for duplicates, which requires comparing each tag against
`tagCopy` (the list of tags) but this list is null on initializing,
and computed properties are calculated before any of the lifecycle hooks

Obviously if you set the min length to show autocomplete to be >0, then
you don't run into this issue because when initializing, the new tag
length is 0, and that gets checked before the results list calculations.